# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Petak:
    properties:
      created_at:
        type: string
      harga:
        type: integer
      id:
        type: string
      jenis_tenda:
        type: string
      kavling_id:
        type: string
      nama:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  model.PetakInput:
    properties:
      harga:
        type: integer
      jenis_tenda:
        type: string
      kavling_id:
        type: string
      nama:
        type: string
      status:
        type: string
    required:
    - harga
    - jenis_tenda
    - kavling_id
    - nama
    - status
    type: object
  model.User:
    properties:
      email:
        description: Email bersifat unik dan tidak boleh kosong
        type: string
      id:
        type: string
      link_ktp:
        description: Link KTP merupakan link ke file KTP
        type: string
      name:
        description: Name tidak boleh kosong
        type: string
      role:
        description: Role merupakan enum yang berisi "admin" dan "user" dan not null
        type: string
    type: object
  model.UserCreate:
    properties:
      city:
        type: string
      confirm_password:
        type: string
      email:
        type: string
      file_ktp:
        type: string
      name:
        type: string
      password:
        type: string
      province:
        type: string
    required:
    - city
    - confirm_password
    - email
    - file_ktp
    - name
    - password
    - province
    type: object
  model.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  model.UserUpdate:
    properties:
      city:
        type: string
      name:
        type: string
      province:
        type: string
    type: object
  utils.ErrorResponseData:
    properties:
      data:
        description: 'Example: nil'
      message:
        description: |-
          The error message.
          Example: "Invalid request"
        type: string
    type: object
  utils.SuccessResponseData:
    properties:
      data:
        description: The actual data returned.
      message:
        description: |-
          The success message.
          Example: "success"
        type: string
    type: object
host: api.perkemahanbedengan.com
info:
  contact:
    email: adityarizky1020@gmail.com
    name: Aditya Rizky Ramadhan
  description: |-
    This is API documentation for Islamind. You can use the API with the following hosts:
    - Production: `api.perkemahanbedengan.com`
  title: Islamind API
  version: "1.0"
paths:
...
  /petak:
    get:
      consumes:
      - application/json
      description: Menampilkan semua data petak
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Petak'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menampilkan semua data petak
      tags:
      - Petak
    post:
      consumes:
      - application/json
      description: Membuat data petak baru
      parameters:
      - description: Data Petak
        in: body
        name: petak
        required: true
        schema:
          $ref: '#/definitions/model.PetakInput'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Petak'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Membuat data petak baru
      tags:
      - Petak
  /petak/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data petak berdasarkan id
      parameters:
      - description: ID Petak
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Petak'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menghapus data petak berdasarkan id
      tags:
      - Petak
    get:
      consumes:
      - application/json
      description: Menampilkan data petak berdasarkan id
      parameters:
      - description: ID Petak
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Petak'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Menampilkan data petak berdasarkan id
      tags:
      - Petak
    put:
      consumes:
      - application/json
      description: Memperbarui data petak berdasarkan id
      parameters:
      - description: ID Petak
        in: path
        name: id
        required: true
        type: string
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data Petak
        in: body
        name: petak
        required: true
        schema:
          $ref: '#/definitions/model.PetakInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.Petak'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Memperbarui data petak berdasarkan id
      tags:
      - Petak
  /user:
    get:
      consumes:
      - application/json
      description: Find user by id
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Find user by id
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user data
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Update user data
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Login user
      tags:
      - User
  /user/logout:
    get:
      consumes:
      - application/json
      description: Logout user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  type: string
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Logout user
      tags:
      - User
  /user/register:
    post:
      consumes:
      - multipart/form-data
      description: Register new user
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseData'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseData'
      summary: Register new user
      tags:
      - User
swagger: "2.0"